requirements.txt
acres==0.3.0
aiofiles==24.1.0
aiohappyeyeballs==2.4.6
aiohttp==3.10.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
asttokens==3.0.0
attrs==25.1.0
beautifulsoup4==4.13.3
certifi==2025.1.31
charset-normalizer==3.4.1
ci-info==0.3.0
click==8.1.8
colorama==0.4.6
comm==0.2.2
configobj==5.0.9
configparser==7.1.0
dataclasses-json==0.6.7
datasets==3.3.2
debugpy==1.8.12
decorator==5.2.1
deepeval==2.4.5
Deprecated==1.2.18
dill==0.3.8
dirtyjson==1.0.8
distro==1.9.0
docx2txt==0.8
etelemetry==0.3.1
execnet==2.1.1
executing==2.2.0
filelock==3.17.0
filetype==1.2.0
fitz==0.0.1.dev2
frontend==0.0.3
frozenlist==1.5.0
fsspec==2024.12.0
googleapis-common-protos==1.68.0
greenlet==3.1.1
grpcio==1.70.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.29.1
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.5.2
iniconfig==2.0.0
ipykernel==6.29.5
ipython==8.32.0
isodate==0.6.1
itsdangerous==2.2.0
jedi==0.19.2
jiter==0.8.2
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jupyter_client==8.6.3
jupyter_core==5.7.2
langchain==0.3.19
langchain-community==0.3.18
langchain-core==0.3.40
langchain-openai==0.3.7
langchain-pinecone==0.2.3
langchain-tests==0.3.12
langchain-text-splitters==0.3.6
langsmith==0.3.11
llama-cloud==0.1.13
llama-cloud-services==0.6.2
llama-index==0.12.20
llama-index-agent-openai==0.4.6
llama-index-cli==0.4.0
llama-index-core==0.12.20
llama-index-embeddings-openai==0.3.1
llama-index-indices-managed-llama-cloud==0.6.8
llama-index-llms-openai==0.3.20
llama-index-multi-modal-llms-openai==0.4.3
llama-index-program-openai==0.3.1
llama-index-question-gen-openai==0.3.0
llama-index-readers-file==0.4.5
llama-index-readers-llama-parse==0.4.0
llama-parse==0.6.2
looseversion==1.3.0
lxml==5.3.1
markdown-it-py==3.0.0
marshmallow==3.26.1
matplotlib-inline==0.1.7
mdurl==0.1.2
multidict==6.1.0
multiprocess==0.70.16
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.4.2
nibabel==5.3.2
nipype==1.9.2
nltk==3.9.1
numpy==1.26.4
openai==1.64.0
opentelemetry-api==1.30.0
opentelemetry-exporter-otlp-proto-common==1.30.0
opentelemetry-exporter-otlp-proto-grpc==1.30.0
opentelemetry-proto==1.30.0
opentelemetry-sdk==1.30.0
opentelemetry-semantic-conventions==0.51b0
orjson==3.10.15
packaging==24.2
pandas==2.2.3
parso==0.8.4
pathlib==1.0.1
pillow==11.1.0
pinecone==5.4.2
pinecone-plugin-inference==3.1.0
pinecone-plugin-interface==0.0.7
platformdirs==4.3.6
pluggy==1.5.0
portalocker==3.1.1
prompt_toolkit==3.0.50
propcache==0.3.0
protobuf==5.29.3
prov==2.0.1
psutil==7.0.0
pure_eval==0.2.3
puremagic==1.28
pyarrow==19.0.1
pydantic==2.10.6
pydantic-settings==2.8.0
pydantic_core==2.27.2
pydot==3.0.4
Pygments==2.19.1
pyparsing==3.2.1
pypdf==5.3.0
pytest==8.3.4
pytest-asyncio==0.25.3
pytest-repeat==0.9.3
pytest-socket==0.7.0
pytest-xdist==3.6.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytils==0.4.1
pytz==2025.1
pywin32==308
pyxnat==1.6.3
PyYAML==6.0.2
pyzmq==26.2.1
rdflib==6.3.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
scipy==1.15.2
sentry-sdk==2.22.0
shellingham==1.5.4
simplejson==3.20.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.38
stack-data==0.6.3
starlette==0.46.0
striprtf==0.0.26
syrupy==4.8.2
tabulate==0.9.0
tenacity==9.0.0
tiktoken==0.9.0
tools==0.1.9
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
traits==7.0.2
typer==0.15.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
wcwidth==0.2.13
wrapt==1.17.2
xxhash==3.5.0
yarl==1.18.3
zipp==3.21.0
zstandard==0.23.0
